v1.3: HSV color picker, fix Color(Long) crash, docs refresh
- Replace predefined color presets with HSV color picker (Hue/Saturation/Brightness sliders)
- Fix crash caused by Color(Long) constructor: convert all Color(0xFFxxxxxx) to Color(0xFFxxxxxx.toInt())
- Fix widget Slider rapid DataStore writes by using local state + onValueChangeFinished
- Rewrite README.md as a user guide
- Update docs (architecture.md, user-guide.md, copilot-instructions.md) to reflect color picker
- Bump version to 1.3.0 (versionCode 4)

+++ app/src/main/java/com/slikharev/shifttrack/MainActivity.kt
@@ -41,11 +41,11 @@ class MainActivity : ComponentActivity() {
             val leaveColor by appDataStore.colorLeave.collectAsState(initial = null)
 
             val colorConfig = ShiftColorConfig(
-                dayColor = dayColor?.let { Color(it.toULong()) } ?: ShiftColors.Day,
-                nightColor = nightColor?.let { Color(it.toULong()) } ?: ShiftColors.Night,
-                restColor = restColor?.let { Color(it.toULong()) } ?: ShiftColors.Rest,
-                offColor = offColor?.let { Color(it.toULong()) } ?: ShiftColors.Off,
-                leaveColor = leaveColor?.let { Color(it.toULong()) } ?: ShiftColors.Leave,
+                dayColor = dayColor?.let { Color(it.toInt()) } ?: ShiftColors.Day,
+                nightColor = nightColor?.let { Color(it.toInt()) } ?: ShiftColors.Night,
+                restColor = restColor?.let { Color(it.toInt()) } ?: ShiftColors.Rest,
+                offColor = offColor?.let { Color(it.toInt()) } ?: ShiftColors.Off,
+                leaveColor = leaveColor?.let { Color(it.toInt()) } ?: ShiftColors.Leave,
             )
 
             ShiftTrackTheme {
