+++ app/src/main/java/com/slikharev/shifttrack/ui/theme/Color.kt
@@ -3,38 +3,38 @@ package com.slikharev.shifttrack.ui.theme
 import androidx.compose.ui.graphics.Color
 
 // Shift type defaults (overridable in settings)
-val ShiftDayDefault = Color(0xFFFFC107)       // Amber
-val ShiftNightDefault = Color(0xFF3F51B5)     // Indigo
-val ShiftRestDefault = Color(0xFF009688)      // Teal
-val ShiftOffDefault = Color(0xFF9E9E9E)       // Grey
-val ShiftOvertimeDefault = Color(0xFFE91E63)  // Pink accent
+val ShiftDayDefault = Color(0xFFFFC107.toInt())       // Amber
+val ShiftNightDefault = Color(0xFF3F51B5.toInt())     // Indigo
+val ShiftRestDefault = Color(0xFF009688.toInt())      // Teal
+val ShiftOffDefault = Color(0xFF9E9E9E.toInt())       // Grey
+val ShiftOvertimeDefault = Color(0xFFE91E63.toInt())  // Pink accent
 
 // Leave type defaults (overridable in settings)
-val LeaveVacationDefault = Color(0xFF4CAF50)  // Green
-val LeaveSickDefault = Color(0xFFF44336)      // Red
-val LeaveStudyDefault = Color(0xFF2196F3)     // Blue
+val LeaveVacationDefault = Color(0xFF4CAF50.toInt())  // Green
+val LeaveSickDefault = Color(0xFFF44336.toInt())      // Red
+val LeaveStudyDefault = Color(0xFF2196F3.toInt())     // Blue
 
 // Material3 seed palette
-val Primary = Color(0xFF1976D2)
-val OnPrimary = Color(0xFFFFFFFF)
-val PrimaryContainer = Color(0xFFD1E4FF)
-val OnPrimaryContainer = Color(0xFF001D36)
-val Secondary = Color(0xFF535F70)
-val OnSecondary = Color(0xFFFFFFFF)
-val SecondaryContainer = Color(0xFFD7E3F7)
-val OnSecondaryContainer = Color(0xFF101C2B)
-val Background = Color(0xFFF8FDFF)
-val Surface = Color(0xFFF8FDFF)
-val SurfaceVariant = Color(0xFFDFE2EB)
-val OnBackground = Color(0xFF191C20)
-val OnSurface = Color(0xFF191C20)
-val OnSurfaceVariant = Color(0xFF43474E)
-val Error = Color(0xFFBA1A1A)
-val OnError = Color(0xFFFFFFFF)
+val Primary = Color(0xFF1976D2.toInt())
+val OnPrimary = Color(0xFFFFFFFF.toInt())
+val PrimaryContainer = Color(0xFFD1E4FF.toInt())
+val OnPrimaryContainer = Color(0xFF001D36.toInt())
+val Secondary = Color(0xFF535F70.toInt())
+val OnSecondary = Color(0xFFFFFFFF.toInt())
+val SecondaryContainer = Color(0xFFD7E3F7.toInt())
+val OnSecondaryContainer = Color(0xFF101C2B.toInt())
+val Background = Color(0xFFF8FDFF.toInt())
+val Surface = Color(0xFFF8FDFF.toInt())
+val SurfaceVariant = Color(0xFFDFE2EB.toInt())
+val OnBackground = Color(0xFF191C20.toInt())
+val OnSurface = Color(0xFF191C20.toInt())
+val OnSurfaceVariant = Color(0xFF43474E.toInt())
+val Error = Color(0xFFBA1A1A.toInt())
+val OnError = Color(0xFFFFFFFF.toInt())
 
-val PrimaryDark = Color(0xFF9ECAFF)
-val OnPrimaryDark = Color(0xFF003258)
-val BackgroundDark = Color(0xFF111318)
-val SurfaceDark = Color(0xFF111318)
-val OnBackgroundDark = Color(0xFFE2E2E9)
-val OnSurfaceDark = Color(0xFFE2E2E9)
+val PrimaryDark = Color(0xFF9ECAFF.toInt())
+val OnPrimaryDark = Color(0xFF003258.toInt())
+val BackgroundDark = Color(0xFF111318.toInt())
+val SurfaceDark = Color(0xFF111318.toInt())
+val OnBackgroundDark = Color(0xFFE2E2E9.toInt())
+val OnSurfaceDark = Color(0xFFE2E2E9.toInt())
